+++ src/components/eye/deployChildren/hardwareDeploy/cameraHardware.vue
@@ -24,8 +24,10 @@
                         style="width: 100%"
                         @selection-change="handleSelectionChange">
                         <el-table-column
-                            type="selection"
                             width="55">
+                            <template slot-scope="scope">
+                                <el-checkbox v-model="scope.row.checked" @change="checked(scope.row.id)" class="checkBoxBtn"></el-checkbox>
+                            </template>
                         </el-table-column>
 
                         <el-table-column
@@ -43,7 +45,9 @@
                         </el-table-column>
                         <el-table-column>
                             <template slot-scope="scope">
-                                <span @click="showPersonDetail(scope.row, '摄像头信息')">编辑</span>
+                                <span @click="showPersonDetail(scope.row, '摄像头信息')">查看</span>
+                                <span @click="fixedInfo(scope.row.id )">编辑</span>
+                                <span @click="deletInfo(scope.row.id)">删除</span>
                             </template>
                         </el-table-column>
                     </el-table>
@@ -104,7 +108,9 @@
             }
         },
         methods:{
-
+            handleSelectionChange(val) {
+                this.multipleSelection = val;
+            },
             addNewInfo(){
                 this.showPersonDetail({},'添加摄像头信息')
                 this.isDisabled=false
@@ -120,6 +126,7 @@
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let cameraObj=[{
+                    typeId:2,
                     id:info.id,
                     name:info.name,
                     positionType:info.positionType,
@@ -140,7 +147,14 @@
                     this.$message.error('修改失败，请稍后再试')
                 })
             },
-            fixedInfo(){
+            fixedInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if(this.choseInfoId.length > 1) {
+                    this.$message.warning('至多选择一条数据')
+                    return
+                }
                 if(this.choseInfoId.length>0){
                     this.cameraList.map((item)=>{
                         if(item.id === this.choseInfoId[0]){
@@ -153,29 +167,47 @@
                     this.$message.error('请选择要修改的摄像头')
                 }
             },
-            deletInfo(){
-                api.camera.deleteCamera(this.choseInfoId).then(res=>{
-                    console.log(res,'删除成功')
-                    for(let i=0;i<this.choseInfoId.length;i++){
-                        this.cameraList=this.cameraList.filter((item,index)=>{
-                            if(item.id===this.choseInfoId[i]){
-                                this.cameraList[index].checked=false
+            deletInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if (this.choseInfoId.length > 0) {
+                    this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        api.camera.deleteCamera(this.choseInfoId).then(res=>{
+                            console.log(res,'删除成功')
+                            for(let i=0;i<this.choseInfoId.length;i++){
+                                this.cameraList=this.cameraList.filter((item,index)=>{
+                                    if(item.id===this.choseInfoId[i]){
+                                        this.cameraList[index].checked=false
+                                    }
+                                    return item.id !== this.choseInfoId[i]
+                                })
                             }
-                            return item.id !== this.choseInfoId[i]
+                            this.$message.success('删除成功')
+                            this.choseInfoId=[]
+                        }).catch(err=>{
+                            console.log(err)
+                            this.$message.error('删除失败，请稍后重试')
                         })
-                    }
-                    this.$message.success('删除成功')
-                    this.choseInfoId=[]
-                }).catch(err=>{
-                    console.log(err)
-                    this.$message.error('删除失败，请稍后重试')
-                })
+                    }).catch(() => {
+                        this.$message.info('取消删除')
+                    })
+
+                } else {
+                    this.$message.error('请选择要删除的数据')
+                }
+
             },
             addNewPerson(info){
                 let index = info.location.includes(',')?info.location.indexOf(','):info.location.indexOf('，')
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let cameraObj=[{
+                    typeId:2,
                     name:info.name,
                     positionType:info.positionType,
                     regionId:info.regionId,
