+++ src/components/eye/deployChildren/hardwareDeploy/gateHardware.vue
@@ -24,33 +24,39 @@
                         style="width: 100%"
                         @selection-change="handleSelectionChange">
                         <el-table-column
-                            type="selection"
                             width="55">
+                            <template slot-scope="scope">
+                                <el-checkbox v-model="scope.row.checked" @change="checked(scope.row.id)" class="checkBoxBtn"></el-checkbox>
+                            </template>
                         </el-table-column>
 
                         <el-table-column
-                            prop="type"
                             label="状态">
+                            <template slot-scope="scope">
+                                <span>{{scope.row.gateType | changeType}}</span>
+                            </template>
                         </el-table-column>
 
                         <el-table-column
-                            prop="area"
+                            prop="regionName"
                             label="所属片区">
                         </el-table-column>
                         <el-table-column
-                            prop="describe"
+                            prop="description"
                             label="摄像头介绍">
                         </el-table-column>
                         <el-table-column>
                             <template slot-scope="scope">
-                                <span @click="showGateDetail(scope.row, 'wifi信息')">编辑</span>
+                                <span @click="showGateDetail(scope.row, '闸机信息')">查看</span>
+                                <span @click="fixedInfo(scope.row.id )">编辑</span>
+                                <span @click="deletInfo(scope.row.id)">删除</span>
                             </template>
                         </el-table-column>
                     </el-table>
 
                     <div class="personInfo" v-for="item in gateList" v-if="isShowGateCard && item.status">
                         <div class="checkBox">
-                            <input type="checkbox" :checked="item.checked" class="checkBtn" @change="checked(item.id)">
+                            <el-checkbox v-model="item.checked" @change="checked(item.id)" class="checkBtn"></el-checkbox>
                         </div>
                         <div class="personType" @click.stop="showGateDetail(item,'闸机信息')">
                             <img src="../../../../../static/img/cameras.png" alt="">
@@ -59,7 +65,7 @@
                                 </span>
                         </div>
                         <div class="specificInfo" >
-                            <p class="area">所属区域：<span>{{item.regionId}}</span></p>
+                            <p class="area">所属区域：<span>{{item.regionName}}</span></p>
                             <p class="type">类&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;型：<span>{{item.gateType | changeType}}</span></p>
                             <p class="sex">描&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;述：<span>{{item.description}}</span></p>
 
@@ -126,6 +132,7 @@
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let gateObj=[{
+                    typeId:3,
                     id:info.id,
                     gateType:info.gateType,
                     name:info.name,
@@ -146,7 +153,14 @@
                     this.$message.error('修改失败，请稍后再试')
                 })
             },
-            fixedInfo(){
+            fixedInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if(this.choseInfoId.length > 1) {
+                    this.$message.warning('至多选择一条数据')
+                    return
+                }
                 if(this.choseInfoId.length>0){
                     this.gateList.map((item)=>{
                         if(item.id === this.choseInfoId[0]){
@@ -159,21 +173,36 @@
                     this.$message.error('请选择要修改的闸机')
                 }
             },
-            deletInfo(){
-                api.gate.deleteGate(this.choseInfoId).then(res=>{
-                    for(let i=0;i<this.choseInfoId.length;i++){
-                        this.gateList=this.gateList.filter((item,index)=>{
-                            if(item.id===this.choseInfoId[i]){
-                                this.gateList[index].checked=false
+            deletInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if (this.choseInfoId.length > 0) {
+                    this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        api.gate.deleteGate(this.choseInfoId).then(res=>{
+                            for(let i=0;i<this.choseInfoId.length;i++){
+                                this.gateList=this.gateList.filter((item,index)=>{
+                                    if(item.id===this.choseInfoId[i]){
+                                        this.gateList[index].checked=false
+                                    }
+                                    return item.id!==this.choseInfoId[i]
+                                })
                             }
-                            return item.id!==this.choseInfoId[i]
+                            this.$message.success('删除成功')
+                            this.choseInfoId=[]
+                        }).catch(err=>{
+                            this.$message.err('删除失败，请稍后重试')
                         })
-                    }
-                    this.$message.success('删除成功')
-                    this.choseInfoId=[]
-                }).catch(err=>{
-                    this.$message.err('删除失败，请稍后重试')
-                })
+                    }).catch(() => {
+                        this.$message.info('取消删除')
+                    })
+                } else {
+                    this.$message.error('请选择要删除的数据')
+                }
 
             },
             addGate(info){
@@ -181,6 +210,7 @@
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let gateObj=[{
+                    typeId:3,
                     gateType:info.gateType,
                     name:info.name,
                     model:info.model,
@@ -207,6 +237,12 @@
                 }
             },
             checked(id){
+                this.gateList = this.gateList.filter(item => {
+                    if (item.id === id) {
+                        item.checked = item.checked
+                    }
+                    return item
+                })
                 console.log(id)
                 if(this.choseInfoId.includes(id)){
                     this.choseInfoId = this.choseInfoId.filter((item)=>{
@@ -354,7 +390,7 @@
                         .checkBtn{
                             position:absolute;
                             right:rem(5);
-                            top:rem(3);
+                            top:rem(-2);
                             cursor:pointer;
                         }
                     }
