+++ src/components/eye/deployChildren/hardwareDeploy/ledHardware.vue
@@ -24,8 +24,10 @@
                         style="width: 100%"
                         @selection-change="handleSelectionChange">
                         <el-table-column
-                            type="selection"
                             width="55">
+                            <template slot-scope="scope">
+                                <el-checkbox v-model="scope.row.checked" @change="checked(scope.row.id)" class="checkBoxBtn"></el-checkbox>
+                            </template>
                         </el-table-column>
                         <el-table-column
                             prop="name"
@@ -33,12 +35,14 @@
                             width="120">
                         </el-table-column>
                         <el-table-column
-                            prop="type"
-                            label="类型">
+                            label="LED类型">
+                            <template slot-scope="scope">
+                                <span>{{scope.row.positionType | changeFilter}}</span>
+                            </template>
                         </el-table-column>
 
                         <el-table-column
-                            prop="regionId"
+                            prop="regionName"
                             label="所属片区">
                         </el-table-column>
 
@@ -48,14 +52,16 @@
                         </el-table-column>
                         <el-table-column>
                             <template slot-scope="scope">
-                                <span @click="showLedDetail(scope.row, 'LED大屏信息')">编辑</span>
+                                <span @click="showLedDetail(scope.row, 'LED大屏信息')">查看</span>
+                                <span @click="fixedInfo(scope.row.id )">编辑</span>
+                                <span @click="deletInfo(scope.row.id)">删除</span>
                             </template>
                         </el-table-column>
                     </el-table>
 
                     <div class="personInfo" v-for="item in ledList" v-if="isShowLedCard && item.status">
                         <div class="checkBox">
-                            <input type="checkbox" :checked="item.checked" class="checkBtn" @change="checked(item.id)">
+                            <el-checkbox v-model="item.checked" @change="checked(item.id)" class="checkBtn"></el-checkbox>
                         </div>
                         <div class="personType" @click.stop="showLedDetail(item,'LED信息')">
                             <img src="../../../../../static/img/cameras.png" alt="">
@@ -64,7 +70,7 @@
                                 </span>
                         </div>
                         <div class="specificInfo" >
-                            <p class="area">所属区域：<span>{{item.regionId}}</span></p>
+                            <p class="area">所属区域：<span>{{item.regionName}}</span></p>
                             <p class="type">LED类型：<span>{{item.positionType | changeFilter}}</span></p>
                             <p class="describe">描&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;述：<span>{{item.description}}</span></p>
 
@@ -112,6 +118,9 @@
             }
         },
         methods:{
+            handleSelectionChange(val) {
+                this.multipleSelection = val;
+            },
             addNewInfo(){
                 this.showLedDetail({},'添加LED大屏信息')
                 this.isDisabled=false
@@ -130,6 +139,7 @@
                 let screenWidth=info.area.substring(0,item)
                 let screenHeight = info.area.substring(item + 1)
                 let ledObj=[{
+                    typeId: 4,
                     id:info.id,
                     positionType:info.positionType,
                     name:info.name,
@@ -150,7 +160,14 @@
                     this.$message.error('修改失败,请稍后再试')
                 })
             },
-            fixedInfo(){
+            fixedInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if(this.choseInfoId.length > 1) {
+                    this.$message.warning('至多选择一条数据')
+                    return
+                }
                 if(this.choseInfoId.length>0){
                     this.ledList.map((item)=>{
                         if(item.id === this.choseInfoId[0]){
@@ -163,21 +180,36 @@
                     this.$message.error('请选择要修改的LED')
                 }
             },
-            deletInfo(){
-                api.led.deleteLed(this.choseInfoId).then(res =>{
-                    for(let i=0;i<this.choseInfoId.length;i++){
-                        this.ledList=this.ledList.filter((item,index)=>{
-                            if(item.id ===this.choseInfoId[i]){
-                                this.ledList[index].checked=false
+            deletInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if (this.choseInfoId.length > 0) {
+                    this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        api.led.deleteLed(this.choseInfoId).then(res =>{
+                            for(let i=0;i<this.choseInfoId.length;i++){
+                                this.ledList=this.ledList.filter((item,index)=>{
+                                    if(item.id ===this.choseInfoId[i]){
+                                        this.ledList[index].checked=false
+                                    }
+                                    return item.id!==this.choseInfoId[i]
+                                })
                             }
-                            return item.id!==this.choseInfoId[i]
+                            this.$message.success('删除成功')
+                            this.choseInfoId=[]
+                        }).catch(err=>{
+                            this.$message.error('删除失败，请稍后再试')
                         })
-                    }
-                    this.$message.success('删除成功')
-                    this.choseInfoId=[]
-                }).catch(err=>{
-                    this.$message.error('删除失败，请稍后再试')
-                })
+                    }).catch(() => {
+                        this.$message.info('取消删除')
+                    })
+                } else {
+                    this.$message.error('请选择要删除的数据')
+                }
             },
             addLed (info){
                 let index = info.location.includes(',')?info.location.indexOf(','):info.location.indexOf('，')
@@ -187,6 +219,7 @@
                 let screenWidth=info.area.substring(0,item)
                 let screenHeight = info.area.substring(item + 1)
                 let ledObj=[{
+                    typeId: 4,
                     positionType:info.positionType,
                     name:info.name,
                     ip:info.ip,
@@ -213,6 +246,12 @@
                 }
             },
             checked(id){
+                this.ledList = this.ledList.filter(item => {
+                    if (item.id === id) {
+                        item.checked = item.checked
+                    }
+                    return item
+                })
                 console.log(id)
                 if(this.choseInfoId.includes(id)){
                     this.choseInfoId = this.choseInfoId.filter((item)=>{
@@ -348,7 +387,7 @@
                         .checkBtn{
                             position:absolute;
                             right:rem(5);
-                            top:rem(3);
+                            top:rem(-2);
                             cursor:pointer;
                         }
                     }
