+++ src/components/eye/deployChildren/hardwareDeploy/monitorsHardware.vue
@@ -24,41 +24,47 @@
                         style="width: 100%"
                         @selection-change="handleSelectionChange">
                         <el-table-column
-                            type="selection"
                             width="55">
+                            <template slot-scope="scope">
+                                <el-checkbox v-model="scope.row.checked" @change="checked(scope.row.id)" class="checkBoxBtn"></el-checkbox>
+                            </template>
                         </el-table-column>
 
-                        <el-table-column
-                            prop="sensorType"
-                            label="类型">
-                        </el-table-column>
+
                         <el-table-column
                             prop="name"
                             label="名称">
                         </el-table-column>
+
                         <el-table-column
                             prop="serialNum"
                             label="设备编号">
                         </el-table-column>
 
                         <el-table-column
-                            prop="regionId"
+                            prop="regionName"
                             label="所属片区">
                         </el-table-column>
                         <el-table-column
                             prop="description"
                             label="描述">
                         </el-table-column>
-                        <el-table-column>
+
+                        <el-table-column label="操作">
+
                             <template slot-scope="scope">
-                                <span @click="showPersonDetail(scope.row, '传感器信息')">编辑</span>
+                                <span @click="showMonitorDetail(scope.row, '传感器信息')">查看</span>
+                                <span @click="fixedInfo(scope.row.id )">编辑</span>
+                                <span @click="deletInfo(scope.row.id)">删除</span>
                             </template>
                         </el-table-column>
                     </el-table>
 
                     <div class="personInfo" v-for="item in monitorsList" v-if="isShowMonitorsCard && item.status">
                         <div class="checkBox">
-                            <input type="checkbox" :checked="item.checked" class="checkBtn" @change="checked(item.id)">
+                            <!--<input type="checkbox" :checked="item.checked" class="checkBtn" @change="checked(item.id)">-->
+                            <el-checkbox v-model="item.checked" @change="checked(item.id)" class="checkBtn"></el-checkbox>
+
                         </div>
                         <div class="personType" @click.stop="showMonitorDetail(item,'传感器信息')">
                             <img src="../../../../../static/img/cameras.png" alt="">
@@ -67,7 +73,7 @@
                                 </span>
                         </div>
                         <div class="specificInfo" >
-                            <p class="name">所属区域：<span>{{item.regionId}}</span></p>
+                            <p class="name">所属区域：<span>{{item.regionName}}</span></p>
                             <p class="sex">描&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;述：<span>{{item.description}}</span></p>
 
                         </div>
@@ -132,6 +138,7 @@
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let monitorsObj =[{
+                    typeId: 6,
                     id:info.id,
                     sensorType:info.sensorType,
                     name:info.name,
@@ -151,7 +158,14 @@
                     this.$message.error('修改失败,请稍后再试')
                 })
             },
-            fixedInfo(){
+            fixedInfo(id){
+                if(id){
+                    this.choseInfoId.push(id)
+                }
+                if(this.choseInfoId.length>1){
+                    this.$message.warning('至多选择一条数据')
+                    return
+                }
                 if(this.choseInfoId.length>0){
                     this.monitorsList.map((item)=>{
                         if(item.id === this.choseInfoId[0]){
@@ -164,27 +178,44 @@
                     this.$message.error('请选择要修改的传感器')
                 }
             },
-            deletInfo(){
-                api.monitor.deleteMonitor(this.choseInfoId).then(res=>{
-                    for(let i=0;i<this.choseInfoId.length;i++){
-                        this.monitorsList=this.monitorsList.filter((item,index)=>{
-                            if(item.id === this.choseInfoId[i]){
-                                this.monitorsList[index].checked=false
+            deletInfo(id){
+                if (id) {
+                    this.choseInfoId.push(id)
+                }
+                if (this.choseInfoId.length > 0) {
+                    this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        api.monitor.deleteMonitor(this.choseInfoId).then(res=>{
+                            for(let i=0;i<this.choseInfoId.length;i++){
+                                this.monitorsList=this.monitorsList.filter((item,index)=>{
+                                    if(item.id === this.choseInfoId[i]){
+                                        this.monitorsList[index].checked=false
+                                    }
+                                    return item.id!==this.choseInfoId[i]
+                                })
                             }
-                            return item.id!==this.choseInfoId[i]
+                            this.$message.success('删除成功')
+                            this.choseInfoId = []
+                        }).catch(err =>{
+                            this.$message.error('删除失败,请稍后重试')
                         })
-                    }
-                    this.$message.success('删除成功')
-                    this.choseInfoId = []
-                }).catch(err =>{
-                    this.$message.error('删除失败,请稍后重试')
-                })
+                    }).catch(() => {
+                        this.$message.info('取消删除')
+                    })
+                } else {
+                    this.$message.error('请选择要删除的数据')
+                }
+
             },
             addMonitors(info){
                 let index = info.location.includes(',')?info.location.indexOf(','):info.location.indexOf('，')
                 let latitude = info.location.substring(0, index)
                 let longitude = info.location.substring(index + 1)
                 let monitorsObj =[{
+                    typeId: 6,
                     sensorType:info.sensorType,
                     name:info.name,
                     manufactor:info.manufactor,
@@ -210,6 +241,12 @@
                 }
             },
             checked(id){
+                this.monitorsList = this.monitorsList.filter(item => {
+                    if (item.id === id) {
+                        item.checked = item.checked
+                    }
+                    return item
+                })
                 console.log(id)
                 if(this.choseInfoId.includes(id)){
                     this.choseInfoId = this.choseInfoId.filter((item)=>{
@@ -222,12 +259,12 @@
             choseType(type){
                 console.log(type)
                 if(type.length===0){
-                    this.choseList=this.monitorsList.filter((item)=>{
+                    this.monitorsList=this.monitorsList.filter((item)=>{
                         item.status=true
                         return item.status === true
                     })
                 }else{
-                    this.choseList=this.monitorsList.filter((item,index)=>{
+                    this.monitorsList=this.monitorsList.filter((item,index)=>{
                         if(type.includes(item.type)){
                             item.status=true
                         }else if(!type.includes(item.type)){
@@ -344,7 +381,7 @@
                         .checkBtn{
                             position:absolute;
                             right:rem(5);
-                            top:rem(3);
+                            top:rem(-2);
                             cursor:pointer;
                         }
                     }
